WIP: embeddings + reduce
